finalized some issues
Fixed the issue where some bf code doesn't work when ran. This would be because i forgot to change the array to 0 when array = 256 and 255 when array = -1.

A problem that could be fixed is the  main array uses ints instead of shorts which would save some data. I probably won't fix this